rename package endpointtimeout to idlenotifier
